refector: changing login and signup forms folders
